Added user routes to seperate file to allow for better organization.
